add an integration test for mindy
Starts up a Pilosa, populates it with data, starts mindy, and then makes a mindy
request and checks the response.

Adds a tiny mindy client as a helper. More invasively, refactors mindy to
separate listening from serving to avoid a race condition in the test.
Technically, just doing `go main.Run()` and then querying mindy is a race since
the server is not guaranteed to be listening if requests are run after that. By
splitting into separate listen and serve methods, we are able to ensure that the
operating system will accept the connection regardless of whether the `go
main.serve()` goroutine has been run yet.